09 - Simple Library/backend/requirements.txt
asgiref==3.7.2
certifi==2023.7.22
charset-normalizer==3.2.0
Django==4.2.4
django-cors-headers==4.2.0
django-environ==0.10.0
django-faker==0.2
django-filter==23.2
django-jazzmin==2.6.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
fake-factory==9999.9.9
Faker==19.2.0
idna==3.4
model-bakery==1.13.0
PyJWT==2.8.0
python-dateutil==2.8.2
pytz==2023.3
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.7.1
urllib3==2.0.4
